requirements.txt
opencv-python == 4.6.0.66
tensorflow == 2.9.1
seaborn == 0.11.2
keras-tuner == 1.1.2
tensorflow-datasets == 4.6.0
streamlit == 1.10.0
idx2numpy == 1.2.3
wandb == 0.12.21
python-dotenv == 0.20.0
pydantic == 1.9.1